% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onemap2pop-data.R
\docType{data}
\name{onemap2pop}
\alias{onemap2pop}
\title{onemap2pop data}
\format{onemap}
\usage{
data(onemap2pop)
}
\description{
Data to be used in the tutorial of composite map
}
\examples{
data(onemap2pop)

}
\references{
Quezada et al. 2018
}
\keyword{datasets}
