% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_2pops.R
\name{rf_2pops}
\alias{rf_2pops}
\title{Multipoint recombination fraction based on two mapping populations}
\usage{
rf_2pops(markers_names, data_P1, data_P2, rftwopoints_P1, rftwopoints_P2,
  LOD = 3, max.rf = 0.5, log10.mintol = -6, max_it = 50)
}
\arguments{
\item{data_P1}{an onemap object class with the information of the first population.}

\item{data_P2}{an onemap object class with the information of the second population.}

\item{rftwopoints_P1}{an rf_2pts object with the information of rf from the first population.}

\item{rftwopoints_P2}{an rf_2pts object with the information of rf from the second population.}

\item{LOD}{minimum LOD Score to declare linkage (defaults to \code{3}).}

\item{max.rf}{maximum recombination fraction to declare linkage (defaults
to \code{0.50}).}

\item{log10.mintol}{log base 10 of the minimum tolerance to converge the EM (defaults to \code{-6})}

\item{max_it}{maximum number of iteration to converge the EM (defaults to \code{50})}

\item{marker_names}{an vector of strings with the marker order.}
}
\value{
A list with two objects of class \code{rf_2pts}, one for each population and the
log-likelihood of the composite  map
}
\description{
The function `rf_2pops` estimates the recombination fraction based on two mapping populations.
It estimates the recombination fractions based on a multipoint approach implemented using the
methodology of Hidden Markov Models (HMM) with the Expectation Maximization (EM) algorithm as
explained in the supplementary material of Quezada et al. (2017).
}
\note{
The thresholds used for \code{LOD} and \code{max.rf} will be used in
subsequent analyses, but can be overriden.
}
\examples{
notrun{
data(onemap2pop)
#Extracting the marker names:
order_LG1POP1 <- colnames(POP1.geno$geno)[LG1_POP1_final$seq.num]
#Computing the rf and likelihood considering information of POP1 and POP2
LG1_POP1order <- rf_2pops(markers_names=order_LG1POP1,
                          data_P1=POP1_geno,
                          data_P2=POP2_geno,
                          rftwopoints_P1 = twopts_POP1,
                          rftwopoints_P2 = twopts_POP2,
                          LOD=3,
                          max.rf=0.5,
                          log10.mintol = -6,
                          max_it = 60)
}

}
\references{
Quezada et al. 2018
}
\author{
Rodrigo R Amadeu \email{rramadeu@gmail.com} and Marianella Quezada \email{marianellaquezada@gmail.com}
}
\keyword{utilities}
