% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ripple_2pops.R
\name{ripple_2pops}
\alias{ripple_2pops}
\title{Compares and displays plausible alternative orders for a given linkage
group based on a composite map}
\usage{
ripple_2pops(markers_names, data_P1, data_P2, twopts_POP1, twopts_POP2,
  LOD = 3, max.rf = 0.5, log10.mintol = -6, max_it = 50,
  window = 3)
}
\arguments{
\item{data_P1}{an onemap object class with the information of the first population.}

\item{data_P2}{an onemap object class with the information of the second population.}

\item{twopts_POP1}{an rf_2pts object with the information of rf from the first population.}

\item{twopts_POP2}{an rf_2pts object with the information of rf from the second population.}

\item{LOD}{minimum LOD Score to declare linkage (defaults to \code{3}).}

\item{max.rf}{maximum recombination fraction to declare linkage (defaults
to \code{0.50}).}

\item{log10.mintol}{log base 10 of the minimum tolerance to converge the EM (defaults to \code{-6})}

\item{max_it}{maximum number of iteration to converge the EM (defaults to \code{50})}

\item{window}{an integer specifying the length of the window size (defaults to 4).}

\item{marker_names}{an vector of strings with the marker order.}
}
\value{
A list with two objects. The first is a matrix with the tested ordered markers,
the second is a vector with the respectivally log-likelihood
}
\description{
For a given sequence of ordered markers, computes the multipoint likelihood
of alternative orders, by shuffling subsets (windows) of markers within the
sequence. For each position of the window, all possible \eqn{(ws)!}{(ws)!}
orders are compared.
}
\details{
Large values for the window size make computations very slow, specially if
there are many partially informative markers.
}
\examples{
notrun{
data(onemap2pop)
#Extracting the marker names:
order_LG1POP1 <- colnames(POP1.geno$geno)[LG1_POP1_final$seq.num]
#RIPPLE algorithm:
LG1_POP1order <- ripple_2pops(markers_names = order_LG1POP1,
                              data_P1 = POP1_geno,
                              data_P2 = POP2_geno,
                              rftwopoints_P1 = twopts_POP1,
                              rftwopoints_P2 = twopts_POP2,
                              LOD = 3,
                              max.rf = 0.5,
                              log10.mintol = -6,
                              max_it = 50,
                              window = 4)
}


}
\references{
Quezada et al. 2018
}
\author{
Rodrigo R Amadeu \email{rramadeu@gmail.com} and Marianella Quezada \email{marianellaquezada@gmail.com}
}
\keyword{utilities}
